Test Scene Updated
Both ships in scene, we now have a sky sphere, the Earth is in the
scene, the sun is in the scene, and the moon is in the scene.
* want to find a way to make the camera follow the ship, I want to see
the ship be well ahead of the camera and in the center of the screen if
possible. If someone can work on that it would be great! I'll work on
adding a few more planets.
* how are we going to access these game objects? They do not have a name
or an id, so how are we going to determine one from another? say for
collisions..