minor change of view
just changed where the camera sees the ship from. In the demo we won't
care much about how precise this is, but later we are going to want to
control the camera differently. For reference look up Star Wars Rogue
Squadron: Rogue Leader and how the camera follows that ship.