Tracing Audio Engine Bug
The bug resides somewhere betwen lines 51 and 69 of Sound.h it is
happening because it is creating sounds every time we want one on the
fly. It is releasing them, but it is not releasing the memory. According
to the documentation System->Release() will release the memory when the
program ends. But I am looking into solutions to this problem.