No More Memory Leak + Optimizations in the code!!
I redesigned how AudioEngine relates to Sound and Stream.
* Sound and Stream have all their memory managed through the AudioSource
Interface. When a user creates an Audio component, they pass in the
filename and the type, AudioSource allows the user access to methods
that are contained within Sound and Stream. Which are inaccessible any
other way.
* Sound and Stream now use a std::pair made of a FMOD::Sound pointer and
a FMOD::Channel pointer, this relationship is core to how sounds are
played through FMOD. Sound has been designed to be used for either
repeated sounds like a rapid fire laser, or for one off sounds.
* Overall memory footprint of AudioEngine, AudioSource, Stream, Sound
and Audio are now significantly reduced due to this streamlining.